+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -17,6 +17,7 @@
 	* Term\Term
 	* Term\TermList
 * Added `Entity::getFingerprint`
+* Added `Entity::setFingerprint`
 
 #### Deprecations